fix 003: cart 필드 이름 변ê경, swagger 라ë¹이브러리 추가

+++ src/main/java/com/mysite/cuffee/order/dto/OrderDto.java
@@ -10,10 +10,10 @@ import lombok.Setter;
 public class OrderDto {
 
     @NotNull(message = "장바구니 ID는 필수입니다.")
-    private int cartId;
+    private Long cartId;
 
     @NotNull(message = "커피 ID는 필수입니다.")
-    private int coffeeId;
+    private Long coffeeId;
 
     @Min(value = 1, message = "수량은 1개 이상이여야 합니다.")
     private int quantity;
